fix prodcut description and rename single product to product description
